add sample preview
