+++ react/src/Components/Views/SamplesView.tsx
@@ -1,14 +1,10 @@
 import React, { Component } from "react";
-import { IBoundingBox } from "../../Interfaces";
 import { Backend } from "../../Utils";
-import { BoundingBoxCanvas } from "../Canvases";
-import { NavBar } from "../UIElements";
+import { NavBar, SamplePreview } from "../UIElements";
 
 
 class SamplesView extends Component<any, any> {
     private params: any;
-    private canvasDOMRect: DOMRect;
-    private canvasStyle: any;
 
     constructor(props: any) {
         super(props);
@@ -20,8 +16,6 @@ class SamplesView extends Component<any, any> {
             },
             samples: [],
         };
-        this.canvasDOMRect = new DOMRect(0, 0, 200, 200);
-        this.canvasStyle = { width: this.canvasDOMRect.width, height: this.canvasDOMRect.height };
     }
 
     public componentDidMount() {
@@ -40,42 +34,9 @@ class SamplesView extends Component<any, any> {
         });
     }
 
-    public boundingBoxPreview(sample: any) {
-        const bb: IBoundingBox = sample.sample;
-        return <div className="tagPreviewOuter" key={sample.media.name}>
-
-            <div className="tagPreviewThumb" style={this.canvasStyle}>
-                <BoundingBoxCanvas
-                    instructionDims={new DOMRect()}
-                    actionDims={new DOMRect()}
-                    viewDims={this.canvasDOMRect}
-                    media_url={sample.media.url}
-                    boundingBox={bb}
-                ></BoundingBoxCanvas>
-            </div>
-
-            <div className="tagPreviewDetails">
-                <div>
-                    <h5>Label</h5>{sample.task.category}
-                </div>
-                <div>
-                    <h5>Confidence</h5> 99%
-                    </div>
-                <div>
-                    <button>✓</button><button style={{ color: "#a81414" }}>✗</button>
-                </div>
-            </div>
-        </div>;
-    }
-
     public render() {
         const samples = this.state.samples.map((sample: any) => {
-            switch (sample.type) {
-                case "BoundingBoxTask":
-                    return this.boundingBoxPreview(sample);
-                default:
-                    return null;
-            }
+            return <SamplePreview sample={sample}></SamplePreview>;
         });
 
         return <div>
